asgards/requirements.txt
pytest
pytest-cov
ruff
azure-devops
msrest
